+++ crates/wasi-common/src/sys/osother.rs
@@ -10,6 +10,11 @@ use std::fs::File;
 use std::io::{self, Read, Write};
 use std::ops::Deref;
 
+pub(crate) trait OsOtherExt {
+    /// Create `OsOther` as `dyn Handle` from null device.
+    fn from_null() -> io::Result<Box<dyn Handle>>;
+}
+
 #[derive(Debug)]
 pub(crate) struct OsOther {
     file_type: Filetype,
