+++ crates/wasi-common/src/sys/windows/osdir.rs
@@ -24,6 +24,10 @@ impl TryFrom<File> for OsDir {
     type Error = io::Error;
 
     fn try_from(file: File) -> io::Result<Self> {
+        let ft = file.metadata()?.file_type();
+        if !ft.is_dir() {
+            return Err(io::Error::from_raw_os_error(libc::EINVAL));
+        }
         let rights = get_rights(&file)?;
         let handle = unsafe { OsHandle::from_raw_handle(file.into_raw_handle()) };
         Self::new(rights, handle)
