+++ crates/wasi-common/src/sys/windows/osfile.rs
@@ -1,9 +1,9 @@
 use super::oshandle::OsHandle;
-use crate::handle::{Handle, HandleRights};
-use crate::sys::osfile::{OsFile, OsFileExt};
+use crate::handle::HandleRights;
+use crate::sys::osfile::OsFile;
 use crate::wasi::{types, RightsExt};
 use std::convert::TryFrom;
-use std::fs::{File, OpenOptions};
+use std::fs::File;
 use std::io;
 use std::os::windows::prelude::{AsRawHandle, FromRawHandle, IntoRawHandle};
 
@@ -11,6 +11,10 @@ impl TryFrom<File> for OsFile {
     type Error = io::Error;
 
     fn try_from(file: File) -> io::Result<Self> {
+        let ft = file.metadata()?.file_type();
+        if !ft.is_file() {
+            return Err(io::Error::from_raw_os_error(libc::EINVAL));
+        }
         let rights = get_rights(&file)?;
         let handle = unsafe { OsHandle::from_raw_handle(file.into_raw_handle()) };
         Ok(Self::new(rights, handle))
@@ -31,12 +35,4 @@ fn get_rights(file: &File) -> io::Result<HandleRights> {
         rights.base |= types::Rights::FD_WRITE;
     }
     Ok(rights)
-}
-
-impl OsFileExt for OsFile {
-    fn from_null() -> io::Result<Box<dyn Handle>> {
-        let file = OpenOptions::new().read(true).write(true).open("NUL")?;
-        let file = Self::try_from(file)?;
-        Ok(Box::new(file))
-    }
 }
