+++ crates/wasi-common/src/sys/windows/osother.rs
@@ -1,10 +1,10 @@
 use super::get_file_type;
 use super::oshandle::OsHandle;
-use crate::handle::HandleRights;
-use crate::sys::osother::OsOther;
+use crate::handle::{Handle, HandleRights};
+use crate::sys::osother::{OsOther, OsOtherExt};
 use crate::wasi::{types, RightsExt};
 use std::convert::TryFrom;
-use std::fs::File;
+use std::fs::{File, OpenOptions};
 use std::io;
 use std::os::windows::prelude::{FromRawHandle, IntoRawHandle};
 
@@ -13,6 +13,9 @@ impl TryFrom<File> for OsOther {
 
     fn try_from(file: File) -> io::Result<Self> {
         let file_type = get_file_type(&file)?;
+        if file_type == types::Filetype::RegularFile || file_type == types::Filetype::Directory {
+            return Err(io::Error::from_raw_os_error(libc::EINVAL));
+        }
         let rights = get_rights(&file_type)?;
         let handle = unsafe { OsHandle::from_raw_handle(file.into_raw_handle()) };
         Ok(Self::new(file_type, rights, handle))
@@ -38,4 +41,12 @@ fn get_rights(file_type: &types::Filetype) -> io::Result<HandleRights> {
     };
     let rights = HandleRights::new(base, inheriting);
     Ok(rights)
+}
+
+impl OsOtherExt for OsOther {
+    fn from_null() -> io::Result<Box<dyn Handle>> {
+        let file = OpenOptions::new().read(true).write(true).open("NUL")?;
+        let file = Self::try_from(file)?;
+        Ok(Box::new(file))
+    }
 }
